// src/app/components/pages-sidebar/codice-sito-gestori/codice-sito-gestori-dto.ts
export class CodiceSitoGestoriDto {
    numcodsito?: any;
    numcodsitoold?: string;
    nomesito?: string;
    gestore?: string;
    tipoimpianto?: string;
    regione?: string;
    provincia?: string;
    comune?: string;
    indirizzo?: string;
    numlocosscem?: string;
    locosscem?: string;
    coordinatelong?: string;
    coordinatelat?: string;

    protocollocheck?: string;
    protcoll?: string;

    linkcondivisia?: string;

    dataprot?: string;
    statoimpianto?: string;
}